// typings/api.d.ts
export type Player = {
  id: number;
  name: string;
  position: 'C' | 'RW' | 'LW' | 'RD' | 'LD';
  league: number;
  team: string;
  season: number;
  gamesPlayed: number;
  timeOnIce: number;
  goals: number;
  assists: number;
  points: number;
  plusMinus: number;
  pim: number;
  ppGoals: number;
  ppAssists: number;
  ppPoints: number;
  ppTimeOnIce: number;
  shGoals: number;
  shAssists: number;
  shPoints: number;
  shTimeOnIce: number;
  fights: number;
  fightWins: number;
  fightLosses: number;
  faceoffs: number | undefined;
  faceoffWins: number | undefined;
  gwg: number | undefined;
  hits: number;
  giveaways: number;
  takeaways: number;
  shotsBlocked: number;
  shotsOnGoal: number;
  gameRating: number;
  offensiveGameRating: number;
  defensiveGameRating: number;
};